requirements.txt
streamlit
landingai_ade
google-generativeai
pandas